Use float64 instead of float32 for metrics of Float type

+++ metrics/processors/node_autoscaling_enricher.go
@@ -56,18 +56,18 @@ func (this *NodeAutoscalingEnricher) Process(batch *core.DataBatch) (*core.DataB
 			memUsed := getInt(metricSet, &core.MetricMemoryUsage)
 
 			if allocatableCpu.MilliValue() != 0 {
-				setFloat(metricSet, &core.MetricNodeCpuUtilization, float32(cpuUsed)/float32(allocatableCpu.MilliValue()))
-				setFloat(metricSet, &core.MetricNodeCpuReservation, float32(cpuRequested)/float32(allocatableCpu.MilliValue()))
+				setFloat(metricSet, &core.MetricNodeCpuUtilization, float64(cpuUsed)/float64(allocatableCpu.MilliValue()))
+				setFloat(metricSet, &core.MetricNodeCpuReservation, float64(cpuRequested)/float64(allocatableCpu.MilliValue()))
 			}
-			setFloat(metricSet, &core.MetricNodeCpuCapacity, float32(capacityCpu.MilliValue()))
-			setFloat(metricSet, &core.MetricNodeCpuAllocatable, float32(allocatableCpu.MilliValue()))
+			setFloat(metricSet, &core.MetricNodeCpuCapacity, float64(capacityCpu.MilliValue()))
+			setFloat(metricSet, &core.MetricNodeCpuAllocatable, float64(allocatableCpu.MilliValue()))
 
 			if allocatableMem.Value() != 0 {
-				setFloat(metricSet, &core.MetricNodeMemoryUtilization, float32(memUsed)/float32(allocatableMem.Value()))
-				setFloat(metricSet, &core.MetricNodeMemoryReservation, float32(memRequested)/float32(allocatableMem.Value()))
+				setFloat(metricSet, &core.MetricNodeMemoryUtilization, float64(memUsed)/float64(allocatableMem.Value()))
+				setFloat(metricSet, &core.MetricNodeMemoryReservation, float64(memRequested)/float64(allocatableMem.Value()))
 			}
-			setFloat(metricSet, &core.MetricNodeMemoryCapacity, float32(capacityMem.Value()))
-			setFloat(metricSet, &core.MetricNodeMemoryAllocatable, float32(allocatableMem.Value()))
+			setFloat(metricSet, &core.MetricNodeMemoryCapacity, float64(capacityMem.Value()))
+			setFloat(metricSet, &core.MetricNodeMemoryAllocatable, float64(allocatableMem.Value()))
 		}
 	}
 	return batch, nil
@@ -80,7 +80,7 @@ func getInt(metricSet *core.MetricSet, metric *core.Metric) int64 {
 	return 0
 }
 
-func setFloat(metricSet *core.MetricSet, metric *core.Metric, value float32) {
+func setFloat(metricSet *core.MetricSet, metric *core.Metric, value float64) {
 	metricSet.MetricValues[metric.MetricDescriptor.Name] = core.MetricValue{
 		MetricType: core.MetricGauge,
 		ValueType:  core.ValueFloat,
